部分代码重构，修改bug
