+++ .gitignore
@@ -8,7 +8,8 @@ yarn-error.log*
 /test/e2e/reports/
 selenium-debug.log
 test/
-/src/components/mypost.vue
+/src/components/HelloWorld.vue
+/src/mock.js
 
 # Editor directories and files
 .idea
