+++ src/App.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-    export default {
+export default {
   name: 'App'
 }
 </script>
